+++ packages/docs/src/components/global/DocRoadmap.vue
@@ -5,12 +5,7 @@
 			rounded
 			class="mb-8"
 		>
-			<VListItem
-				href="https://github.com/orgs/assurance-maladie-digital/projects/2/views/5"
-				target="_blank"
-				rel="noopener noreferrer"
-				class="px-4"
-			>
+			<VListItem class="px-4">
 				<VListItemAvatar class="d-none d-sm-inline-flex">
 					<VIcon
 						color="primary"
@@ -26,13 +21,36 @@
 					</VListItemTitle>
 
 					<VListItemSubtitle>
-						La version 2 du Design System est en cours de finalisation, les demandes d'ajout de composants ou de fonctionnalités ne sont plus acceptées, mais nous continuerons les corrections de bugs jusqu'à fin 2023.
+						La version 2 du Design System est en maintenance jusqu’à fin 2023.
 					</VListItemSubtitle>
 				</VListItemContent>
+			</VListItem>
+		</VSheet>
 
-				<VIcon class="ml-4">
-					{{ newTabIcon }}
-				</VIcon>
+		<VSheet
+			outlined
+			rounded
+			class="mb-8"
+		>
+			<VListItem class="px-4">
+				<VListItemAvatar class="d-none d-sm-inline-flex">
+					<VIcon
+						color="primary"
+						x-large
+					>
+						{{ tagIcon }}
+					</VIcon>
+				</VListItemAvatar>
+
+				<VListItemContent>
+					<VListItemTitle class="mb-1">
+						Version 2.0.0-bridge.x
+					</VListItemTitle>
+
+					<VListItemSubtitle>
+						La version Bridge du Design System est en cours de développement et permettra de migrer sur Vue 3 sans changements sur les composants.
+					</VListItemSubtitle>
+				</VListItemContent>
 			</VListItem>
 		</VSheet>
 
@@ -61,7 +79,7 @@
 					</VListItemTitle>
 
 					<VListItemSubtitle>
-						La version 3 du Design System est en cours de conception et son développement commencera en Juin 2023.
+						La version 3 du Design System est en cours de conception et son développement commencera fin 2023.
 					</VListItemSubtitle>
 				</VListItemContent>