+++ packages/docs/src/data/examples/select-btn-field/usage.vue
@@ -1,0 +1,34 @@
+<template>
+	<SelectBtnField
+		v-model="value"
+		:items="items"
+		label="Moyen de contact"
+	/>
+</template>
+
+<script lang="ts">
+	import Vue from 'vue';
+	import Component from 'vue-class-component';
+
+	import { SelectBtnItem } from '@cnamts/vue-dot/src/patterns/SelectBtnField/types';
+
+	@Component
+	export default class SelectBtnFieldUsage extends Vue {
+		value: string[] = [];
+
+		items: SelectBtnItem[] = [
+			{
+				text: 'Email',
+				value: 'email'
+			},
+			{
+				text: 'Courrier',
+				value: 'courrier'
+			},
+			{
+				text: 'SMS',
+				value: 'sms'
+			}
+		];
+	}
+</script>