+++ packages/form-builder/src/components/FormField/mixins/fieldMap.ts
@@ -8,10 +8,10 @@ export interface IFieldMap {
 // We import them all because the form can use any of them
 import DateField from '../fields/DateField.vue';
 import NumberField from '../fields/NumberField.vue';
-import PasswordField from '../fields/PasswordField.vue';
+import FormPasswordField from '../fields/FormPasswordField.vue';
 import ChoicePeriodField from '../fields/ChoicePeriodField.vue';
 import ChoiceField from '../fields/ChoiceField.vue';
-import RangeField from '../fields/RangeField.vue';
+import FormRangeField from '../fields/FormRangeField.vue';
 import TextareaField from '../fields/TextareaField.vue';
 import TextField from '../fields/TextField.vue';
 
@@ -19,10 +19,10 @@ import TextField from '../fields/TextField.vue';
 	components: {
 		DateField,
 		NumberField,
-		PasswordField,
+		FormPasswordField,
 		ChoicePeriodField,
 		ChoiceField,
-		RangeField,
+		FormRangeField,
 		TextareaField,
 		TextField
 	}
@@ -31,10 +31,10 @@ export class FieldMap extends Vue {
 	fieldMap: IFieldMap = {
 		date: 'DateField',
 		number: 'NumberField',
-		password: 'PasswordField',
+		password: 'FormPasswordField',
 		period: 'ChoicePeriodField',
 		select: 'ChoiceField',
-		range: 'RangeField',
+		range: 'FormRangeField',
 		text: 'TextField',
 		textarea: 'TextareaField'
 	};