+++ packages/vue-dot/src/patterns/PasswordField/config.ts
@@ -1,0 +1,6 @@
+export const config = {
+	validateOnBlur: true,
+	btn: {
+		icon: true
+	}
+};