+++ packages/vue-dot/src/patterns/RangeField/config.ts
@@ -1,0 +1,10 @@
+export const config = {
+	textField: {
+		hideDetails: true,
+		class: 'ma-3'
+	},
+	rangeSlider: {
+		hideDetails: true,
+		class: 'align-center mt-2 mb-6'
+	}
+};