+++ packages/vue-dot/src/patterns/index.ts
@@ -1,3 +1,4 @@
+import AlertWrapper from './AlertWrapper';
 import ContextualMenu from './ContextualMenu';
 import CookieBanner from './CookieBanner';
 import DataListGroup from './DataListGroup';
@@ -13,15 +14,19 @@ import LangBtn from './LangBtn';
 import NirField from './NirField';
 import NotificationBar from './NotificationBar';
 import PaginatedTable from './PaginatedTable';
+import PasswordField from './PasswordField';
 import PeriodField from './PeriodField';
 import PhoneField from './PhoneField';
+import RangeField from './RangeField';
 import RatingPicker from './RatingPicker';
 import SearchListField from './SearchListField';
+import SelectBtnField from './SelectBtnField';
 import SubHeader from './SubHeader';
 import TableToolbar from './TableToolbar';
 import UploadWorkflow from './UploadWorkflow';
 
 export const patterns = {
+	AlertWrapper,
 	ContextualMenu,
 	CookieBanner,
 	DataListGroup,
@@ -37,10 +42,13 @@ export const patterns = {
 	NirField,
 	NotificationBar,
 	PaginatedTable,
+	PasswordField,
 	PeriodField,
 	PhoneField,
+	RangeField,
 	RatingPicker,
 	SearchListField,
+	SelectBtnField,
 	SubHeader,
 	TableToolbar,
 	UploadWorkflow